fix: restore demo route support on Vercel
