+++ src/app/layout/sidenav/sidenav.component.html
@@ -19,4 +19,23 @@
       </a>
     </div>
   </ng-container>
+  <div class="sidenav__group">
+    <div class="sidenav__group-title">Демо</div>
+    <a
+      matRipple
+      routerLink="/demo"
+      class="sidenav__item"
+      routerLinkActive="active"
+    >
+      <span>Демо</span>
+    </a>
+    <a
+      matRipple
+      routerLink="/demo/drivers"
+      class="sidenav__item"
+      routerLinkActive="active"
+    >
+      <span>Демо — Водители</span>
+    </a>
+  </div>
 </nav>
