+++ vercel.json
@@ -1,4 +1,17 @@
 {
   "version": 2,
-  "installCommand": "npm ci --loglevel=error || npm install --no-audit --no-fund --loglevel=error"
+  "installCommand": "npm ci --loglevel=error || npm install --no-audit --no-fund --loglevel=error",
+  "builds": [
+    {
+      "src": "package.json",
+      "use": "@vercel/static-build",
+      "config": { "distDir": "dist/g-track-frontend/browser" }
+    }
+  ],
+  "routes": [
+    { "src": "/assets/(.*)", "dest": "/assets/$1" },
+    { "src": "/favicon\\.ico", "dest": "/favicon.ico" },
+    { "src": "/robots\\.txt", "dest": "/robots.txt" },
+    { "src": "/.*", "dest": "/index.html" }
+  ]
 }
